refactor: add necessary things without breaking code
